refactor: extract pure logic into parse and report modules
Split the single-file action into:
- parse.ts: turn 'git apply --check' stderr into a conflict file list,
  applying excluded path patterns
- report.ts: build the check-run report markdown from conflicts
- conflibot.ts: the orchestrator class (GitHub API + git operations)
- index.ts: entry point

The pure modules take plain data instead of Octokit response objects,
so they can be unit-tested without mocking the GitHub API. No behavior
change.
